Updated codestyle
