+++ tests/src/Validator/Rule/ArrayInputTest.php
@@ -19,7 +19,7 @@ use Bluz\Validator\Rule\ArrayInput;
 class ArrayInputTest extends Tests\TestCase
 {
     /**
-     * @var \Bluz\Validator\Rule\StringRule
+     * @var \Bluz\Validator\Rule\StringInput
      */
     protected $validator;
