+++ tests/src/Validator/ValidatorTest.php
@@ -32,6 +32,7 @@ class ValidatorTest extends Tests\TestCase
      */
     public function testStaticCallsShouldReturnNewValidator()
     {
+        $this->assertInstanceOf('Bluz\Validator\Validator', Validator::arrayInput());
         $this->assertInstanceOf('Bluz\Validator\Validator', Validator::string());
         $this->assertInstanceOf('Bluz\Validator\Validator', Validator::notEmpty());
     }
